moved calq to genie/quality/
