+++ src/CMakeLists.txt
@@ -1,3 +1,2 @@
 add_subdirectory(apps)
 add_subdirectory(genie)
-add_subdirectory(calq)
