+++ src/genie/CMakeLists.txt
@@ -16,4 +16,5 @@ add_subdirectory(module)
 add_subdirectory(util)
 add_subdirectory(quality/qvwriteout)
 add_subdirectory(name/tokenizer)
+add_subdirectory(quality/calq)
 add_subdirectory(quality/qvcalq)
