+++ src/genie/quality/calq/calq_coder.h
@@ -12,7 +12,7 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/log.h"
+#include "log.h"
 
 // -----------------------------------------------------------------------------
