+++ src/genie/quality/calq/circular_buffer.h
