+++ src/genie/quality/calq/error_exception_reporter.h
@@ -10,7 +10,7 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/exceptions.h"
+#include "exceptions.h"
 
 // -----------------------------------------------------------------------------
