+++ src/genie/quality/calq/filter_buffer.cc
@@ -1,4 +1,4 @@
-#include "calq/filter_buffer.h"
+#include "filter_buffer.h"
 
 // -----------------------------------------------------------------------------
 
@@ -10,7 +10,7 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/error_exception_reporter.h"
+#include "error_exception_reporter.h"
 
 // -----------------------------------------------------------------------------
