+++ src/genie/quality/calq/filter_buffer.h
@@ -9,7 +9,7 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/circular_buffer.h"
+#include "circular_buffer.h"
 
 // -----------------------------------------------------------------------------
