+++ src/genie/quality/calq/genotyper.cc
@@ -1,4 +1,4 @@
-#include "calq/genotyper.h"
+#include "genotyper.h"
 
 // -----------------------------------------------------------------------------
 
@@ -13,8 +13,8 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/error_exception_reporter.h"
-#include "calq/log.h"
+#include "error_exception_reporter.h"
+#include "log.h"
 
 // -----------------------------------------------------------------------------
