+++ src/genie/quality/calq/lloyd_max_quantizer.cc
@@ -1,4 +1,4 @@
-#include "calq/lloyd_max_quantizer.h"
+#include "lloyd_max_quantizer.h"
 
 // -----------------------------------------------------------------------------
