+++ src/genie/quality/calq/lloyd_max_quantizer.h
@@ -11,9 +11,9 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/exceptions.h"
-#include "calq/probability_distribution.h"
-#include "calq/quantizer.h"
+#include "exceptions.h"
+#include "probability_distribution.h"
+#include "quantizer.h"
 
 // -----------------------------------------------------------------------------
