+++ src/genie/quality/calq/log.cc
@@ -1,4 +1,4 @@
-#include "calq/log.h"
+#include "log.h"
 
 #include <memory>
