+++ src/genie/quality/calq/qual_decoder.cc
@@ -1,10 +1,10 @@
-#include "calq/qual_decoder.h"
+#include "qual_decoder.h"
 
 // -----------------------------------------------------------------------------
 
-#include "calq/calq_coder.h"
-#include "calq/error_exception_reporter.h"
-#include "calq/qual_encoder.h"
+#include "calq_coder.h"
+#include "error_exception_reporter.h"
+#include "qual_encoder.h"
 
 // -----------------------------------------------------------------------------
