+++ src/genie/quality/calq/qual_decoder.h
@@ -9,7 +9,7 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/quantizer.h"
+#include "quantizer.h"
 
 // -----------------------------------------------------------------------------
