+++ src/genie/quality/calq/qual_encoder.cc
@@ -1,4 +1,4 @@
-#include "calq/qual_encoder.h"
+#include "qual_encoder.h"
 
 // -----------------------------------------------------------------------------
 
@@ -8,8 +8,8 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/calq_coder.h"
-#include "calq/error_exception_reporter.h"
+#include "calq_coder.h"
+#include "error_exception_reporter.h"
 
 // -----------------------------------------------------------------------------
