+++ src/genie/quality/calq/quantizer.cc
@@ -1,4 +1,4 @@
-#include "calq/quantizer.h"
+#include "quantizer.h"
 
 // -----------------------------------------------------------------------------
 
@@ -7,8 +7,8 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/error_exception_reporter.h"
-#include "calq/log.h"
+#include "error_exception_reporter.h"
+#include "log.h"
 
 // -----------------------------------------------------------------------------
