+++ src/genie/quality/calq/sam_pileup_deque.h
@@ -7,7 +7,7 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/sam_pileup.h"
+#include "sam_pileup.h"
 
 // -----------------------------------------------------------------------------
