+++ src/genie/quality/calq/softclip_spreader.h
@@ -12,7 +12,7 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/circular_buffer.h"
+#include "circular_buffer.h"
 
 // -----------------------------------------------------------------------------
