+++ src/genie/quality/calq/uniform_min_max_quantizer.h
@@ -1,9 +1,9 @@
-#ifndef CALQ_UNIFORM_QUANTIZER_H_
-#define CALQ_UNIFORM_QUANTIZER_H_
+#ifndef CALQ_UNIFORM_MIN_MAX_QUANTIZER_H_
+#define CALQ_UNIFORM_MIN_MAX_QUANTIZER_H_
 
 // -----------------------------------------------------------------------------
 
-#include "calq/quantizer.h"
+#include "uniform_quantizer.h"
 
 // -----------------------------------------------------------------------------
 
@@ -11,14 +11,15 @@ namespace calq {
 
 // -----------------------------------------------------------------------------
 
-class UniformQuantizer : public Quantizer
+class UniformMinMaxQuantizer : public UniformQuantizer
 {
  public:
-    UniformQuantizer(const int& valueMin,
-                     const int& valueMax,
-                     const int& nrSteps
+    UniformMinMaxQuantizer(const int& valueMin,
+                           const int& valueMax,
+                           const int& nrSteps
     );
-    ~UniformQuantizer() override;
+
+    ~UniformMinMaxQuantizer() override;
 };
 
 // -----------------------------------------------------------------------------
@@ -27,7 +28,7 @@ class UniformQuantizer : public Quantizer
 
 // -----------------------------------------------------------------------------
 
-#endif  // CALQ_UNIFORM_QUANTIZER_H_
+#endif  // CALQ_UNIFORM_MIN_MAX_QUANTIZER_H_
 
 // -----------------------------------------------------------------------------
 // -----------------------------------------------------------------------------
