+++ src/genie/quality/calq/uniform_quantizer.cc
@@ -1,4 +1,4 @@
-#include "calq/uniform_quantizer.h"
+#include "uniform_quantizer.h"
 
 // -----------------------------------------------------------------------------
 
@@ -11,7 +11,7 @@
 
 // -----------------------------------------------------------------------------
 
-#include "calq/error_exception_reporter.h"
+#include "error_exception_reporter.h"
 
 // -----------------------------------------------------------------------------
